refactor(*): convert to typescript
use typescript for the whole project
