--- lib/inquirer.js
@@ -1,36 +1,0 @@
-const inquirer = require("inquirer");
-const files = require("./files");
-
-module.exports = {
-  askProjectConfig: () => {
-    const argv = require("minimist")(process.argv.slice(2));
-    const questions = [
-      {
-        name: "userName",
-        type: "input",
-        message: "Enter your github username or organization",
-        validate: function (value) {
-          if (value.length) {
-            return true;
-          } else {
-            return "Please enter  your username or organization";
-          }
-        },
-      },
-      {
-        name: "packageName",
-        type: "input",
-        message: "Enter a name for your package (no camel case)",
-        default: argv._[0] || files.getCurrentDirectoryBase(),
-        validate: function (value) {
-          if (value.length) {
-            return true;
-          } else {
-            return "Please enter a name for your package";
-          }
-        },
-      },
-    ];
-    return inquirer.prompt(questions);
-  },
-};
